.FinalTask: add addRecordToHospitalizationStatsTest()

+++ FinalTask/src/test/java/ru/croc/javaschool/finaltask/database/DatabaseHandlerTest.java
@@ -2,6 +2,7 @@ package ru.croc.javaschool.finaltask.database;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
+import ru.croc.javaschool.finaltask.object.HospitalizationsReport;
 import ru.croc.javaschool.finaltask.object.InfectionsReport;
 
 import java.sql.Connection;
@@ -32,23 +33,49 @@ public class DatabaseHandlerTest {
         var actualLastReport = new InfectionsReport();
 
         // adding new record
-        handler.addRecordToInfectionStats(expectedReport);
-
-        // reading last record in table
-        Connection connection = DriverManager.getConnection(handler.url);
-        Statement statement = connection.createStatement();
-        var resultSet = statement.executeQuery("SELECT date, infection_cases_count, recovery_cases_count " +
-                "FROM Infection_stats");
-        while (resultSet.next()) {
-            actualLastReport = new InfectionsReport(
-                    LocalDate.parse(resultSet.getString(1)),
-                    resultSet.getInt(2),
-                    resultSet.getInt(3));
+        if (handler.addRecordToInfectionStats(expectedReport)) {
+            // reading last record in table
+            Connection connection = DriverManager.getConnection(handler.url);
+            Statement statement = connection.createStatement();
+            var resultSet = statement.executeQuery("SELECT date, infection_cases_count, recovery_cases_count " +
+                    "FROM Infection_stats");
+            while (resultSet.next()) {
+                actualLastReport = new InfectionsReport(
+                        LocalDate.parse(resultSet.getString(1)),
+                        resultSet.getInt(2),
+                        resultSet.getInt(3));
+            }
+            // deleting added record
+            statement.executeUpdate("DELETE FROM Infection_stats WHERE date = '" + expectedReport.getDate().toString() + "'");
+            connection.close();
+            statement.close();
         }
-        // deleting added record
-        statement.executeUpdate("DELETE FROM Infection_stats WHERE date = '" + expectedReport.getDate().toString() + "'");
-        connection.close();
-        statement.close();
+
+        Assertions.assertEquals(expectedReport, actualLastReport);
+    }
+
+    @Test
+    public void addRecordToHospitalizationStatsTest() throws SQLException {
+        var expectedReport = new HospitalizationsReport(LocalDate.now(), 15, 20);
+        var actualLastReport = new HospitalizationsReport();
+
+        // adding new record
+        if (handler.addRecordToHospitalizationStats(expectedReport)) {
+            // reading last record in table
+            Connection connection = DriverManager.getConnection(handler.url);
+            Statement statement = connection.createStatement();
+            var resultSet = statement.executeQuery("SELECT * FROM Hospitalization_stats");
+            while (resultSet.next()) {
+                actualLastReport = new HospitalizationsReport(
+                        LocalDate.parse(resultSet.getString(1)),
+                        resultSet.getInt(2),
+                        resultSet.getInt(3));
+            }
+            // deleting added record
+            statement.executeUpdate("DELETE FROM Hospitalization_stats WHERE date = '" + expectedReport.getDate().toString() + "'");
+            connection.close();
+            statement.close();
+        }
 
         Assertions.assertEquals(expectedReport, actualLastReport);
     }
